Removed hGraphQL from profile
