+++ Projects/Profile/Sources/Views/Screens/DeleteAccount/DeleteRequestLoadingView.swift
@@ -3,7 +3,6 @@ import PresentableStore
 import SwiftUI
 import hCore
 import hCoreUI
-import hGraphQL
 
 public struct DeleteRequestLoadingView: View {
     @PresentableStore var store: ProfileStore
