+++ Projects/Profile/Sources/Views/Screens/SettingsScreen/SettingsView.swift
@@ -4,7 +4,6 @@ import PresentableStore
 import SwiftUI
 import hCore
 import hCoreUI
-import hGraphQL
 
 struct SettingsView: View {
     @PresentableStore var store: ProfileStore
